TwoStageDectection/requirements.txt
jittor==1.3.9.14
numpy==2.3.0
Pillow==11.2.1
tqdm==4.66.2
